requirements.txt
beautifulsoup4==4.6.0
jellyfish==0.6.1
Jinja2==2.10
marshmallow==2.15.4
marshmallow-mongoengine==0.9.1
mongoengine==0.15.3
newspaper3k==0.2.6
pymongo==3.6.0
python-decouple==3.1
redis==2.10.6
requests==2.18.4
rq==0.11.0
tldextract==2.2.0
tornado==5.0.2
tweepy==3.6.0
urllib3==1.22
blinker==1.4
apispec==0.39.0
factory_boy==2.11.1